refactor: storage options for local & session
